refactor: unify processscanner and process/watcher into processlifecycle
Combine two separate process-handling packages into a single
adapters/inbound/agents/processlifecycle package that owns the full
process lifecycle: birth detection (polling) + death detection (kqueue).

Deduplicates shared OS utilities (pgrep, lsof, CWD resolution) that
were independently implemented in both packages.

Closes #56
