+++ core/adapters/inbound/agents/processlifecycle/osutil.go
@@ -1,0 +1,65 @@
+// Package processlifecycle owns the full process lifecycle for agent sessions:
+// birth detection (polling) and death detection (kqueue). It unifies the
+// previously separate processscanner and process/watcher packages, deduplicating
+// shared OS utilities (pgrep, lsof, CWD resolution).
+package processlifecycle
+
+import (
+	"context"
+	"fmt"
+	"os/exec"
+	"strconv"
+	"strings"
+	"time"
+)
+
+// FindProcesses returns PIDs of processes whose name exactly matches name.
+func FindProcesses(name string) ([]int, error) {
+	ctx, cancel := context.WithTimeout(context.Background(), 2*time.Second)
+	defer cancel()
+	out, err := exec.CommandContext(ctx, "pgrep", "-x", name).Output()
+	if err != nil {
+		// pgrep exits 1 when there are no matches — not an error.
+		if exit, ok := err.(*exec.ExitError); ok && exit.ExitCode() == 1 {
+			return nil, nil
+		}
+		return nil, err
+	}
+	var pids []int
+	for _, line := range strings.Split(strings.TrimSpace(string(out)), "\n") {
+		line = strings.TrimSpace(line)
+		if line == "" {
+			continue
+		}
+		pid, err := strconv.Atoi(line)
+		if err == nil && pid > 0 {
+			pids = append(pids, pid)
+		}
+	}
+	return pids, nil
+}
+
+// ProcessCWD returns the working directory of pid using lsof.
+func ProcessCWD(pid int) (string, error) {
+	ctx, cancel := context.WithTimeout(context.Background(), 2*time.Second)
+	defer cancel()
+	out, err := exec.CommandContext(ctx, "lsof", "-a", "-p", strconv.Itoa(pid), "-d", "cwd", "-Fn").Output()
+	if err != nil {
+		return "", fmt.Errorf("lsof cwd pid %d: %w", pid, err)
+	}
+	for _, line := range strings.Split(string(out), "\n") {
+		if strings.HasPrefix(line, "n") {
+			return strings.TrimPrefix(line, "n"), nil
+		}
+	}
+	return "", fmt.Errorf("cwd not found for pid %d", pid)
+}
+
+// CWDToProjectDir converts a working directory path to the directory name used
+// by Claude Code under ~/.claude/projects/. Claude Code replaces both "/" and
+// "." with "-", so "/Users/ingo/projects/foo" becomes "-Users-ingo-projects-foo"
+// and "/path/.hidden/sub" becomes "-path--hidden-sub".
+func CWDToProjectDir(cwd string) string {
+	s := strings.ReplaceAll(cwd, "/", "-")
+	return strings.ReplaceAll(s, ".", "-")
+}
